s3_vector/requirements.txt
# AWS SDK and S3 Vector support
boto3>=1.35.0
botocore>=1.35.0

# LangChain and AWS integration
langchain>=0.1.0
langchain-core>=0.1.0
langchain-aws>=0.1.0
langchain-community>=0.0.1



# Machine Learning and Embeddings
transformers>=4.36.0
sentence-transformers>=2.2.2
torch>=2.1.0
numpy>=1.24.0

# Multi-modal capabilities (lightweight)
pillow>=10.0.0
# Note: CLIP models are downloaded automatically by sentence-transformers
# No additional OpenAI API access required

# Data handling and utilities
pandas>=2.0.0
dataclasses-json>=0.6.0

# Progress and visualization
tqdm>=4.66.0
tabulate>=0.9.0

# Development and testing
python-dotenv>=1.0.0

# Jupyter notebook support
jupyter>=1.0.0
jupyterlab>=4.0.0
matplotlib>=3.7.0
ipywidgets>=8.0.0